requirements.txt
load_dotenv
controller
aiosmtpd